requirements.txt
streamlit==1.28.1
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai>=0.8.3
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4